+++ GEMFILE
@@ -1,13 +1,12 @@
 source 'https://rubygems.org/'
 
-ruby '2.2.0'
-
 gem 'data_mapper'
 gem 'sinatra'
 gem 'dm-postgres-adapter'
 gem 'bcrypt'
 gem 'rack-flash3'
 gem 'rake'
+gem 'sinatra-partial'
 
 group :test do
   gem 'rspec'
