+++ app/controllers/application.rb
@@ -1,0 +1,4 @@
+get '/' do
+  @tweets = Tweet.all
+  erb :index
+end
