+++ app/server.rb
@@ -1,10 +1,15 @@
 require 'sinatra'
+require 'sinatra/partial'
 require 'data_mapper'
 require 'rack-flash'
 require './app/lib/tweet'
 require './app/lib/user'
 require_relative 'helpers/application'
 require_relative 'data_mapper_setup'
+require_relative 'controllers/application'
+require_relative 'controllers/tweets'
+require_relative 'controllers/sessions'
+require_relative 'controllers/users'
 
 set :public_folder, Proc.new { File.join(root, "..", "public") }
 
@@ -12,57 +17,3 @@ enable :sessions
 set :session_secret, 'super awesome secret password'
 use Rack::Flash
 use Rack::MethodOverride
-
-get '/' do
-  @tweets = Tweet.all
-  erb :index
-end
-
-post '/tweets' do
-  message = params["message"]
-  user = params["user"]
-  Tweet.create(:message => message, :user => current_user.name)
-  redirect to('/')
-end
-
-get '/users/new' do
-  @user = User.new
-  erb :"users/new"
-end
-
-post '/users' do
-  @user = User.create(:email => params[:email],
-                     :password => params[:password],
-                     :password_confirmation => params[:password_confirmation],
-                     :name => params[:name],
-                     :username => params[:username])
-    if @user.save
-      session[:user_id] = @user.id
-      redirect to('/')
-    else
-      flash.now[:errors] = @user.errors.full_messages
-      erb :"users/new"
-    end
-end
-
-get '/sessions/new' do
-  erb :"sessions/new"
-end
-
-post '/sessions' do
-  email, password = params[:email], params[:password]
-  user = User.authenticate(email, password)
-  if user
-    session[:user_id] = user.id
-    redirect to('/')
-  else
-    flash[:errors] = ["The email or password is incorrect"]
-    erb :"sessions/new"
-  end
-end
-
-delete '/sessions' do
-  session.clear
-  redirect "/"
-  flash.now[:notice] = "You've been log-out!"
-end
